+++ src/DIXTableView.h
@@ -18,7 +18,7 @@
 @end
 
 @interface NSObject(DIXTableViewDelegate)
-- (NSMenu*) tableView: (NSTableView *) tableView menuForTableColumn: (NSTableColumn*) column row: (int) row;
+- (NSMenu*) tableView: (NSTableView *) tableView menuForTableColumn: (NSTableColumn*) column row: (NSInteger) row;
 	//delegate will be asked what menu to show (if not implemented by delegate [self menu] is used)
 
 //- (NSDragOperation) draggingSourceOperationMaskForLocal:(BOOL)isLocal;
