+++ src/MainWindowController.m
@@ -543,7 +543,7 @@
 	}
 	
 	//remove any supernumerary menu items (removed all items if is there is no app which can open this file)
-	unsigned removeMenuItemsFromIndex = ([apps defaultAppDesc] != nil) ? [[apps additionalAppDescs] count] +2 : 0;
+	NSUInteger removeMenuItemsFromIndex = ([apps defaultAppDesc] != nil) ? [[apps additionalAppDescs] count] +2 : 0;
 	
 	while ( ((unsigned) [_openWithSubMenu numberOfItems]) > removeMenuItemsFromIndex )
 		[_openWithSubMenu removeItemAtIndex: [_openWithSubMenu numberOfItems] -1];
